Add 400 response on MultiParser exceptions [2]
