+++ starlette/exceptions.py
@@ -1,30 +1,11 @@
 import asyncio
-import http
 import typing
 
 from starlette.concurrency import run_in_threadpool
-from starlette.formparsers import ParserException
+from starlette.http_exception import HTTPException
 from starlette.requests import Request
 from starlette.responses import PlainTextResponse, Response
 from starlette.types import ASGIApp, Message, Receive, Scope, Send
-
-
-class HTTPException(Exception):
-    def __init__(
-        self,
-        status_code: int,
-        detail: typing.Optional[str] = None,
-        headers: typing.Optional[dict] = None,
-    ) -> None:
-        if detail is None:
-            detail = http.HTTPStatus(status_code).phrase
-        self.status_code = status_code
-        self.detail = detail
-        self.headers = headers
-
-    def __repr__(self) -> str:
-        class_name = self.__class__.__name__
-        return f"{class_name}(status_code={self.status_code!r}, detail={self.detail!r})"
 
 
 class ExceptionMiddleware:
@@ -41,10 +22,7 @@ class ExceptionMiddleware:
         self._status_handlers: typing.Dict[int, typing.Callable] = {}
         self._exception_handlers: typing.Dict[
             typing.Type[Exception], typing.Callable
-        ] = {
-            HTTPException: self.http_exception,
-            ParserException: self._parser_exception,
-        }
+        ] = {HTTPException: self.http_exception}
         if handlers is not None:
             for key, value in handlers.items():
                 self.add_exception_handler(key, value)
@@ -113,6 +91,3 @@ class ExceptionMiddleware:
         return PlainTextResponse(
             exc.detail, status_code=exc.status_code, headers=exc.headers
         )
-
-    def _parser_exception(self, request: Request, exc: ParserException) -> Response:
-        return PlainTextResponse(exc.message, 400)
