+++ tests/test_formparsers.py
@@ -1,11 +1,10 @@
 import os
 import typing
 
-import pytest
-
 from starlette.formparsers import UploadFile, _user_safe_decode
 from starlette.requests import Request
 from starlette.responses import JSONResponse
+from starlette.testclient import TestClient
 
 
 class ForceMultipartDict(dict):
@@ -390,16 +389,19 @@ def test_user_safe_decode_ignores_wrong_charset():
     assert result == "abc"
 
 
-def test_missing_boundary_parameter(test_client_factory):
+def test_missing_boundary_parameter(
+    test_client_factory: typing.Callable[..., TestClient]
+):
     client = test_client_factory(app)
-    with pytest.raises(KeyError, match="boundary"):
-        client.post(
-            "/",
-            data=(
-                # file
-                b'Content-Disposition: form-data; name="file"; filename="\xe6\x96\x87\xe6\x9b\xb8.txt"\r\n'  # noqa: E501
-                b"Content-Type: text/plain\r\n\r\n"
-                b"<file content>\r\n"
-            ),
-            headers={"Content-Type": "multipart/form-data; charset=utf-8"},
-        )
+    res = client.post(
+        "/",
+        data=(
+            # file
+            b'Content-Disposition: form-data; name="file"; filename="\xe6\x96\x87\xe6\x9b\xb8.txt"\r\n'  # noqa: E501
+            b"Content-Type: text/plain\r\n\r\n"
+            b"<file content>\r\n"
+        ),
+        headers={"Content-Type": "multipart/form-data; charset=utf-8"},
+    )
+    assert res.status_code == 400
+    assert res.text == "Missing boundary in multipart."
